remove moka dependency, fix formats

+++ contracts/test/HypNative.t.sol
@@ -12,6 +12,7 @@ import "../src/mocks/MockERC721.sol";
 
 contract HypNativeTest is Test, HyperlaneAddressesConfig {
     using TypeCasts for address;
+
     uint256 public sourceChain;
     uint256 public destinationChain;
     uint32 public destinationChainId = uint32(31338);
@@ -19,19 +20,20 @@ contract HypNativeTest is Test, HyperlaneAddressesConfig {
     address public deployer = 0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266;
     address public recipient = address(1);
     address public hypNativeTokenAddress = 0x09635F643e140090A9A8Dcd712eD6285858ceBef;
+
     function setUp() public {
         sourceChain = vm.createFork(vm.rpcUrl("source"));
         destinationChain = vm.createFork(vm.rpcUrl("destination"));
     }
 
-     receive() external payable {}
+    receive() external payable {}
 
     /**
      * @dev Test checks that source chain part send native hyp tokens succeed
      */
     function testXChainSendNativeTokensSourcePart() public {
         uint256 payGasFees = 0.001 ether;
-        uint amount = 0.2 ether;
+        uint256 amount = 0.2 ether;
         vm.selectFork(sourceChain);
         HypNative hypNativeToken = HypNative(payable(hypNativeTokenAddress));
         vm.deal(deployer, 1 ether);
@@ -39,7 +41,9 @@ contract HypNativeTest is Test, HyperlaneAddressesConfig {
         uint256 lockedNativeAssetsBefore = hypNativeToken.balanceOf(address(hypNativeToken));
 
         vm.prank(deployer);
-        hypNativeToken.transferRemote{value: payGasFees + amount}(destinationChainId, recipient.addressToBytes32(), amount);
+        hypNativeToken.transferRemote{value: payGasFees + amount}(
+            destinationChainId, recipient.addressToBytes32(), amount
+        );
 
         assertEq(hypNativeToken.balanceOf(address(hypNativeToken)), lockedNativeAssetsBefore + amount);
     }
@@ -48,15 +52,13 @@ contract HypNativeTest is Test, HyperlaneAddressesConfig {
      * @dev Test checks that source chain part send native hyp tokens succeed
      */
     function testXChainSendNativeFailWhenSendWithoutGasFees() public {
-        uint amount = 0.2 ether;
+        uint256 amount = 0.2 ether;
         vm.selectFork(sourceChain);
         HypNative hypNativeToken = HypNative(payable(hypNativeTokenAddress));
         vm.deal(deployer, 1 ether);
 
-
         vm.prank(deployer);
         vm.expectRevert(bytes("IGP: insufficient interchain gas payment"));
         hypNativeToken.transferRemote{value: amount}(destinationChainId, recipient.addressToBytes32(), amount);
-
     }
 }
